Update mms.py to download DES and fix FGM
(1) Old FGM version number was hardcoded into fgm_survey.  This is not sustainable (survey data in March 2016 has *4* distinct versions, because they're absolute maniacs) so instead we ask SDC for a JSON with filenames matching those dates and fetch accordingly.  Processing may change down the line for some reason I haven't foreseen, but at least this way data are downloading...

(2) The FPI team is slightly less absolutely maniacal so the version changes can be handled more gracefully - or just left as is.  The DES and DIS can be grabbed with a one-letter difference so I've gone and done that.